ref(node): move config back to node package - the depency cycle issue will be resolved in another way
