+++ node/node.go
@@ -12,7 +12,6 @@ import (
 	"go.uber.org/fx"
 
 	"github.com/celestiaorg/celestia-node/core"
-	"github.com/celestiaorg/celestia-node/node/config"
 	"github.com/celestiaorg/celestia-node/service/block"
 )
 
@@ -25,7 +24,7 @@ var log = logging.Logger("node")
 // * Light
 type Node struct {
 	Type   Type
-	Config *config.Config
+	Config *Config
 
 	// the Node keeps a reference to the DI App that controls the lifecycles of services registered on the Node.
 	app *fx.App
